+++ RedPandaIDE/translations/RedPandaIDE_pt_BR.ts
@@ -1433,7 +1433,7 @@
     </message>
     <message>
         <source>Max Undo Steps</source>
-        <translation>Quantidade máxima de passos a serem desfeitos</translation>
+        <translation type="vanished">Quantidade máxima de passos a serem desfeitos</translation>
     </message>
     <message>
         <source>Default file encoding</source>
@@ -1464,19 +1464,39 @@
         <translation>UTF-8 BOM</translation>
     </message>
     <message>
-        <source>Max Undo Memory Usage</source>
-        <translation type="unfinished"></translation>
-    </message>
-    <message>
         <source>MB</source>
         <translation type="unfinished">MB</translation>
     </message>
     <message>
-        <source>Auto reformat code before saving files</source>
+        <source>Parse TODOs</source>
         <translation type="unfinished"></translation>
     </message>
     <message>
-        <source>Parse TODOs</source>
+        <source>Memory Usage</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Steps</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Action before saving files</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Reformat Code</source>
+        <translation type="unfinished">Reformatar código</translation>
+    </message>
+    <message>
+        <source>Remove Trailing Spaces</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>None</source>
+        <translation type="unfinished"></translation>
+    </message>
+    <message>
+        <source>Limits for Undo</source>
         <translation type="unfinished"></translation>
     </message>
 </context>
@@ -6946,6 +6966,10 @@
     <message>
         <source>Command: %1 %2</source>
         <translation>Comando: %1 %2</translation>
+    </message>
+    <message>
+        <source>Compiling...</source>
+        <translation type="unfinished"></translation>
     </message>
 </context>
 <context>